move _ClientImpl to its own file
